requirements.txt
llama-index
llama-hub
unstructured
openai
nltk
python-dotenv
sentence-transformers
torch
transformers
chainlit
numpy
packaging
llama-index-core
requests
beautifulsoup4
selenium
pillow
PyPDF2
python-docx
pandas
openpyxl
moviepy
librosa
discord.py
langdetect
PyNaCl